[WIP] getting reading working (again??)

+++ Fabric.Tests/DataTests.cs
@@ -17,23 +17,21 @@ namespace Fabric.Tests {
             }
         }
 
-        public class TestSubPage : DataPage {
-            public TestSubPage(string name) : base(name) { }
-        }
-
         private JsonSerializerSettings _basicSerializerSettings;
 
         private readonly string _basicRootPageJson =
                 "{\"Name\":\"root\",\"SchemaName\":null,\"PageData\":null,\"Children\":{\"none\":[\"TestItem1\",\"TestItem2\"]}}"
             ;
 
-        private FabricDatabase CreateTestingDb() {
+        private FabricDatabase CreateTestingDb(bool deleteOld = true) {
             var databaseDir = Path.Combine(Directory.GetCurrentDirectory(), "TestingDB");
 
-            if (Directory.Exists(databaseDir))
-                Directory.Delete(databaseDir, true);
+            if (deleteOld) {
+                if (Directory.Exists(databaseDir))
+                    Directory.Delete(databaseDir, true);
 
-            Directory.CreateDirectory(databaseDir);
+                Directory.CreateDirectory(databaseDir);
+            }
 
             var database = new FabricDatabase(databaseDir);
             database.Initialise();
@@ -83,7 +81,7 @@ namespace Fabric.Tests {
         }
 
         [Fact]
-        public void RootPageChild_ShouldDelete_Correctly()
+        public void RootPageChild_ShouldLoad_Correctly()
         {
             var database = CreateTestingDb();
 
@@ -92,6 +90,47 @@ namespace Fabric.Tests {
 
             Assert.True(File.Exists(Path.Combine(database.DatabaseRoot, "none", "TestItem1", "dataPage.json")));
             Assert.True(File.Exists(Path.Combine(database.DatabaseRoot, "none", "TestItem2", "dataPage.json")));
+
+            // clear database reference
+            database = null;
+            database = CreateTestingDb(false);
+
+            var item = database.Root.GetChild("TestItem1");
+        }
+
+        [Fact]
+        public void RootPageChild_ShouldUpdateCorrectly()
+        {
+            var database = CreateTestingDb();
+            database.Root.AddChild("TestItem1", "none", "{}");
+
+            Assert.True(File.Exists(Path.Combine(database.DatabaseRoot, "none", "TestItem1", "dataPage.json")));
+
+            var child = database.Root.GetChild("TestItem1");
+
+            Assert.Equal("{}", child.PageData);
+
+            var data = JsonConvert.SerializeObject(new {test = "TestObject"});
+            child.PageData = data;
+            child.SaveChanges();
+
+            var text = File.ReadAllText(Path.Combine(database.DatabaseRoot, "none", "TestItem1", "dataPage.json"));
+            var node = JsonConvert.DeserializeObject<DataPage>(text, database.SerializerSettings);
+
+            var actual = JsonUtils.Uglify(JsonUtils.RemoveProperty(node.PageData, "modifiedTimestamp", true));
+            var expected = JsonUtils.Uglify(JsonUtils.RemoveProperty(data, "modifiedTimestamp", true));
+
+            Assert.Equal(expected, actual, true);
+        }
+
+        [Fact]
+        public void RootPageChild_ShouldDelete_Correctly()
+        {
+            var database = CreateTestingDb();
+
+            database.Root.AddChild("TestItem1", "none", "{}");
+
+            Assert.True(File.Exists(Path.Combine(database.DatabaseRoot, "none", "TestItem1", "dataPage.json")));
 
             database.Root.DeleteChild("TestItem1", "none");
